https://youtu.be/hjEE6c0oOI0
You can use multiple variable elements inside the data element. Each variable element describes a property that may be set on the layout to be used in binding expressions within the layout file. This commit is for learning more about Variables from Data Binding library and how to use the viewmodel and its observables/methods .